Oil Platform layout and vehicle blueprints
Added the basic layout of the Oil Platform level. It is currently set up in a sort of "demo" mode. All spawn points are filled with crafting stations and all required resources are scattered around the map. The recipes for everything is found floating above various parts of the map. The balloon escape vehicle now functions similarly to the crafting stations. Further improvements are required. The buttons set the current recipe that the balloon will try to make.