First part of mobile escape vehicle
Once all the components of the escape vehicle have been built, it will now spawn another identical looking actor in the same position and then delete itself. This actor sets the timer to 30 seconds, triggers a flag and can be boarded by players who jump in its collision box. (Most of this is untested. Beware of bugs)