Grab Bag of stuff
Put everything in the Workshop level into folders. Added a fourth option to destructible floors: None, which prevents them from breaking at all. Replaced middle floor in Oil rig level with destructible floors (Note all are set to None currently and the texture is not correct). Fixed the escape vehicle on the Oil Rig level. Fixed the moving background objects on the Oil Rig level. Fixed a self pin that escaped the purge.